Refactor User Interface elements from Add and List to UI class
